WEB-4025: replaced old dashboard with tasks, removed unused components
