+++ app/(candidate)/dashboard/components/AnimatedProgressBar.js
